major refactoring AND CSS styling
