+++ client/package.json
@@ -14,6 +14,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "clsx": "^1.1.1",
     "material-ui-search-bar": "^1.0.0-beta.14",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
