+++ client/src/App.js
@@ -9,6 +9,7 @@ import { fab } from '@fortawesome/free-brands-svg-icons';
 import { faPaperPlane } from '@fortawesome/free-solid-svg-icons';
 import Home from './components/Home';
 import Form from './components/Form/Form';
+import Inbox from './components/Inbox/Inbox';
 import List from './components/List';
 import NavBar from './components/NavBar/NavBar';
 import { businesses, volunteers } from './constant';
