+++ client/src/components/inbox/components/index.js
@@ -1,0 +1,13 @@
+import Card from './Card';
+import Input from './Input';
+import Sidebar from './Sidebar';
+import { Text, Title, Subtitle } from './Text';
+
+export {
+    Card,
+    Input,
+    Sidebar,
+    Text,
+    Title,
+    Subtitle,
+}
