+++ client/src/modules/inbox.module.css
@@ -1,0 +1,57 @@
+.root {
+    display: flex;
+    height: 100%;
+}
+
+.container {
+    background: #fafafa;
+    border-radius: 5px;
+}
+
+.inbox {
+    padding: 10px;
+    flex: 1;
+}
+
+.sidebar {
+    width: 200px;
+    padding: 10px 0;
+    margin-right: 5px;
+    font-size: small;
+}
+
+.messages {
+    height: calc(100% - 55px);
+    overflow-y: scroll;
+    overscroll-behavior-y: contain;
+    scroll-snap-type: y mandatory;
+}
+
+.fill {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+    justify-content: center;
+}
+
+.img {
+    display: block;
+    margin: 30px auto;
+    border-radius: 50%;
+}
+
+.card {
+    border-radius: 5px;
+    cursor: pointer;
+    padding: 5px;
+}
+
+.card:last-child {
+    scroll-snap-align: end;
+}
+
+.chat {
+    box-sizing: border-box;
+    height: 100%;
+    padding: 10px;
+}
